add missed favicon images + fix removing account

+++ src/hooks/screen/myProfile/useMyProfile.js
@@ -2,6 +2,7 @@ import { useEffect, useState } from "react";
 import { removeAccount } from "../../../redux/updateMyProfile/removeAccount/thunk/removeAccountThunk";
 import { logoutUser } from "../../../redux/logout/thunk/logoutThunk";
 import { removeAccountClearState } from "../../../redux/updateMyProfile/removeAccount/thunk/removeAccountThunk";
+import { disconnectIOSocket } from "../../../components/navbar/disconnectIOSocket";
 import { useSelector, useDispatch } from "react-redux";
 
 const useMyProfile = () => {
@@ -10,6 +11,7 @@ const useMyProfile = () => {
 
   const { userData } = useSelector((state) => state.login);
   const { accountRemoved } = useSelector((state) => state.isAccountRemoved);
+  const { socket } = useSelector((state) => state.socketData);
 
   const handleTrash = (e) => {
     e.preventDefault();
@@ -27,11 +29,14 @@ const useMyProfile = () => {
   };
 
   useEffect(() => {
-    if (accountRemoved) {
-      dispatch(removeAccountClearState());
-      dispatch(logoutUser(userData._id, userData.email));
-    }
-  }, [accountRemoved, userData._id, userData.email, dispatch]);
+    (async () => {
+      if (accountRemoved && socket.connected) {
+        dispatch(removeAccountClearState());
+        await disconnectIOSocket(socket, userData._id);
+        dispatch(logoutUser(userData._id, userData.email));
+      }
+    })();
+  }, [accountRemoved, socket, userData._id, userData.email, dispatch]);
 
   return { showRemoveAccount, handleTrash, handleCancel, handleDeleteAccount };
 };
